+++ changelog.rst
@@ -10,6 +10,13 @@ Changelog
 -   [d]—docs improvement
 
 
+2.2.5 (March 23, 2021)
+======================
+
+-   [+] Add ``onDelete`` pragma that lets add ``ON DELETE`` constraints to foreign keys.
+-   [r] Improve formatting of import statements: use ``std/`` notation.
+
+
 2.2.4 (March 21, 2021)
 ======================